day-03-invoice-generator/requirements.txt
reportlab>=4.0
pytest>=7.0
pdfminer.six